+++ include/init.hpp
@@ -127,6 +127,17 @@ std::string add_squo(const std::string &str)
 {
     return "'"+str+"'";
 }
+std::string replace_all(std::string str,const std::string &from,const std::string &to)
+{
+    size_t pos=0;
+    while((pos=str.find(from,pos))!=std::string::npos)
+    {
+        str.replace(pos,from.length(),to);
+        pos+=to.length();
+    }
+    return str;
+}
+
 
 // local path
 namespace Init